+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/type/Call.java
@@ -36,9 +36,6 @@ public class Call {
     private String id;
     private List<DetectPoint> detectPoints;
 
-    private List<Integer> sourceComponentIDs;
-    private List<Integer> targetComponentIDs;
-
     public Call() {
         sourceComponents = new ArrayList<>();
         targetComponents = new ArrayList<>();