+++ oap-server/server-storage-plugin/storage-elasticsearch-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/elasticsearch/base/ColumnTypeEsMapping.java
@@ -23,20 +23,19 @@ import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.List;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
-import org.apache.skywalking.oap.server.core.storage.model.DataTypeMapping;
+import org.apache.skywalking.oap.server.core.storage.model.ElasticSearchExtension;
 import org.apache.skywalking.oap.server.core.storage.type.StorageDataComplexObject;
 
-public class ColumnTypeEsMapping implements DataTypeMapping {
+public class ColumnTypeEsMapping {
 
-    @Override
-    public String transform(Class<?> type, Type genericType) {
+    public String transform(Class<?> type, Type genericType, final ElasticSearchExtension elasticSearchExtension) {
         if (Integer.class.equals(type) || int.class.equals(type) || Layer.class.equals(type)) {
             return "integer";
         } else if (Long.class.equals(type) || long.class.equals(type)) {
             return "long";
         } else if (Double.class.equals(type) || double.class.equals(type)) {
             return "double";
-        } else if (String.class.equals(type)) {
+        } else if (String.class.equals(type) || elasticSearchExtension.isKeyword()) {
             return "keyword";
         } else if (StorageDataComplexObject.class.isAssignableFrom(type)) {
             return "text";
@@ -46,7 +45,7 @@ public class ColumnTypeEsMapping implements DataTypeMapping {
             return "text";
         } else if (List.class.isAssignableFrom(type)) {
             final Type elementType = ((ParameterizedType) genericType).getActualTypeArguments()[0];
-            return transform((Class<?>) elementType, elementType);
+            return transform((Class<?>) elementType, elementType, elasticSearchExtension);
         } else {
             throw new IllegalArgumentException("Unsupported data type: " + type.getName());
         }