+++ docs/en/changes/changes.md
@@ -41,6 +41,9 @@
 * Use TTL-driven interval settings for the `measure-default` group in BanyanDB.
 * Fix wrong group of non time-relative metadata in BanyanDB.
 * Refactor `StorageData#id` to the new StorageID object from a String type.
+* Support multiple component IDs in the service topology level.
+* Add `ElasticSearch.Keyword` annotation to declare the target field  type as `keyword`.
+* [Breaking Change] Field `component_id` of `service_relation_client_side` and `service_relation_server_side` have been replaced by `component_ids`(keyword).
 
 #### UI